+++ playerLib/src/main/java/com/zj/playerLib/upstream/RawResourceDataSource.java
@@ -1,8 +1,3 @@
-//
-// Source code recreated from a .class file by IntelliJ IDEA
-// (powered by FernFlower decompiler)
-//
-
 package com.zj.playerLib.upstream;
 
 import android.content.Context;
@@ -93,7 +88,7 @@ public final class RawResourceDataSource extends BaseDataSource {
         } else {
             int bytesRead;
             try {
-                int bytesToRead = this.bytesRemaining == -1L ? readLength : (int)Math.min(this.bytesRemaining, (long)readLength);
+                int bytesToRead = this.bytesRemaining == -1L ? readLength : (int)Math.min(this.bytesRemaining, readLength);
                 bytesRead = this.inputStream.read(buffer, offset, bytesToRead);
             } catch (IOException var6) {
                 throw new RawResourceDataSourceException(var6);
@@ -107,7 +102,7 @@ public final class RawResourceDataSource extends BaseDataSource {
                 }
             } else {
                 if (this.bytesRemaining != -1L) {
-                    this.bytesRemaining -= (long)bytesRead;
+                    this.bytesRemaining -= bytesRead;
                 }
 
                 this.bytesTransferred(bytesRead);
