+++ playerLib/src/main/java/com/zj/playerLib/upstream/crypto/CryptoUtil.java
@@ -1,8 +1,3 @@
-//
-// Source code recreated from a .class file by IntelliJ IDEA
-// (powered by FernFlower decompiler)
-//
-
 package com.zj.playerLib.upstream.crypto;
 
 final class CryptoUtil {
@@ -16,7 +11,7 @@ final class CryptoUtil {
             long hash = 0L;
 
             for(int i = 0; i < input.length(); ++i) {
-                hash ^= (long)input.charAt(i);
+                hash ^= input.charAt(i);
                 hash += (hash << 1) + (hash << 4) + (hash << 5) + (hash << 7) + (hash << 8) + (hash << 40);
             }
